Reducing top-ten-percent (true values only), making it searchable
Adding top-ten-percent to the Reception section label. I've prototyped various ways to add it - including labels and didn't like the results when it's visible in every context. With that in mind - having it as a searchable property and visible in a subtle way on a product card might work better

Refactoring additional reductions into a shared func with a shared redux